+++ src/components/profile-form/profile-form.module.css
@@ -1,0 +1,10 @@
+.profile__form {
+  display: flex;
+  flex-direction: column;
+  gap: 24px
+}
+
+.profile__changeButtons {
+  display: flex;
+  justify-content: flex-end;
+}
